MB-39791: Move modules around to ease converting Query to go modules
query (the only client of go-curl) uses go.mod to import go-curl from
couchbasedeps, but we also need to map it to that directory in repo sync
so that non-modules projects that depend on query can find go-curl on
GOPATH. We temporarily map it to both old and new locations to make it
possible for CV to test the remaining changes while not breaking builds.

Also, move gocbcore-transactions and n1fty under goproj rather than
godeps. This won't affect current builds since both paths are on GOPATH,
but it makes the replace directives in query/go.mod and n1fty/go.mod
much shorter, which in turn makes it easier to emulate for people who
build query but don't use "repo sync" to get the source.

Change-Id: I859bbb90236f2eca56a4a9b3ed6eb7de87bf19bd
Reviewed-on: http://review.couchbase.org/c/manifest/+/137328
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>