MB-41868: Bump HdrHistogram revision to fix int64->int32 conversion bug
hdr_iter takes the total_count of the histogram (h) at construction.
The count used in the histogram is a 64 bit count. When we have more
than std::numeric_limits<int32_t>::max() counts in the histogram the
total_count field of the hdr_iter will underflow when we create the
iterator and various >0 checks will check in and prevent iteration of
the histogram.

Change-Id: I688959a261f4de6fef72738a19c6280a2f018889
Reviewed-on: http://review.couchbase.org/c/manifest/+/137733
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>