MB-38476: Checkout spdlog v1.8.5 (latest) to third_party/spdlog_new
We want to upgrade spdlog to v1.4.0 or newer to use their static
library mode in the future. Upgrade to v1.8.5 as it's the latest
version.

The v1.1.0 and v1.8.5 spdlog APIs are not compatible so checkout this
version of spdlog to spdlog_new alongside the current spdlog checkout
so that we can test this using CV with linked tlm and kv_engine
changes. In a future patch we can remove/rename spdlog_new such that
we only include one spdlog checkout.

Change-Id: I90e0ee3b979f11b770810a2dadc3b90cffc5ba61
Reviewed-on: http://review.couchbase.org/c/manifest/+/154428
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>