Introducing Elixir builds; Morpheus -> 8.0.0
Per PM decision, "Elixir" will be 7.2.0, which is the version number
formerly associated with Morpheus. To keep build numbers meaningful,
the first Elixir build will be 7.2.0-1200, which is slightly higher
than the most recent Morpheus 7.2.0 build. Morpheus is changed to
8.0.0 and will start building at 8.0.0-1000.

Elixir builds will be hourly as most engineering effort is focused
there. Morpheus and branch-master are redundant for the time being, so
we set the branch-master build interval to weekly. Morpheus builds are
set to daily as no Morpheus-specific activity is expected in the near
future (or even possible, since all projects are reading from their
master/main branches in both manifests).

Also removed "mobile-service" from branch-master which was an oversight
from some time ago.

Change-Id: Ib85b545b7368bca90c23312223cfda825bd5ffc5
Reviewed-on: https://review.couchbase.org/c/manifest/+/174911
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>