MB-54258: Do not build googletest and benchmark with -Werror
Clang 13 (shipped with macOS 12) introduces some new warnings which
caused the build of benchmark and googletest to fail, as these
dependencies compile with -Werror "turn warnings into errors".

This patch updates the CC manifest to our patched versions of these
dependencies, which disable -Werror entirely. Note that we don't use 
-Werror to build any of our projects, so this makes the build process for these dependencies more consistent with everything else.

Change-Id: I6163d0fc21f8e8692162f8291786771971d1e4da
Reviewed-on: https://review.couchbase.org/c/manifest/+/182005
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>