MB-54376: Do not build benchmark with -Werror for MH
Clang 13 (shipped with macOS 12) introduces some new warnings which
caused the build of benchmark to fail, as it compiles with -Werror
"turn warnings into errors".

This patch updates the MH manifest to our patched version of benchmark,
which disables -Werror entirely. Note that we don't use -Werror to
build any of our projects, so this makes the build process for these
dependencies more consistent with everything else.

Change-Id: I0ab81ed99cb6817a5f051032d2454fe1fd7d54b6
Reviewed-on: https://review.couchbase.org/c/manifest/+/182223
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>