MB-55204: product-texts on 'master' git branch for all active versions
Per PM, all releases should get the newest available license text,
including older point releases (eg. 7.0.6, 6.6.7, etc) and even MP
releases. So pointing it to 'master' for all versions.

I've also proposed a change to build-from-manifest which will prevent
product-texts changes from triggering builds. Therefore we don't need to
worry about a bunch of unwanted builds of old versions being triggered
whenever the license or readme text is updated.

I've verified that this won't cause unexpected changes. In particular,
the locked SHA 7ca2324 that many manifests were pointing to is the same
SHA as the current tip of both the 'master' and 'neo' branches. The
'cheshire-cat' branch is a few commits behind, but not diverged. The
commits used by couchbase-tableau-connector manifests were between the
'cheshire-cat' and 'neo' commits, and also not diverged from master. So
this change should strictly update any future builds of those product
versions to the latest license and readme.

Change-Id: Ia0c01a1c460c3ae43a9a62d1b4f795baec049d67
Reviewed-on: https://review.couchbase.org/c/manifest/+/186190
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>